Split driver.cs into three files
