+++ objcgen/objcgen.csproj
@@ -324,6 +324,8 @@
     <Compile Include="NameGenerator.cs" />
     <Compile Include="processor.cs" />
     <Compile Include="system-check.cs" />
+    <Compile Include="embedder.cs" />
+    <Compile Include="utils.cs" />
   </ItemGroup>
   <ItemGroup>
     <Folder Include="support\" />
